+++ .changeset/3ff4268c/changes.md
@@ -1,1 +1,1 @@
-- add onCommentPermalinkClicked prop so consumers can handle that click for scrolling or other purpose
+- add onCommentPermalinkClick prop so consumers can handle that click for scrolling or other purpose